Fixed The ADMIN and USER programs
Used a lot of Linus's code in order to Actually merge the two codes. Everything is now working correctly! Good luck on the merge!

There are 2 new methods:

runAdminProgram
runUserProgram

Also edited a lot of other stuff with minor changes such as renaming

Added a "Test" user in the generated list